// src/app/models/employee.models.ts
export class Employee {
    education: string | undefined;
    institutename(institutename: any) {
      throw new Error('Method not implemented.');
    }
    id?: number = 0;
    firstname: string = '';
    lastname: string = '';
    birthdate: string = '';
    gender: string = '';
    address: string = '';
    contactno: number = 0;
    pancard: string = '';
    bloodgroup: string = '';
    emergencycontactname: string = ('');
    emergencycontactno: number = 0;
    maritalstatus: string = ('');
    hobbies: string = ('');
    familydetails: string = ('');
    instituename: string = ('');
    degree: string = ('');
    startdate: number = 0;
    enddate: number = 0;
    companyname: string = ('');
    designation: string = ('');
    joiningdate: number = 0;
    revealingdate: number = 0;
}